refactor(editor): extract EffectHandler for side-effect interpretation
Move apply_effects/2, all apply_effect/2 clauses, and supporting helpers
(apply_log_effect, maybe_spawn_prettify, spawn_prettify_task) into
MingaEditor.Handlers.EffectHandler. The editor keeps a delegating type
alias and defdelegate for backward compatibility.

Promotes recover_swap_entries, restore_session, maybe_refresh_tool_picker,
reset_nav_flash_tracking, cancel_nav_flash, cancel_yank_flash, and
maybe_schedule_warning_popup to @doc false public helpers. Removes the
duplicate cancel_flash_if_active and execute_flash_effects wrappers.

Editor reduced from 1,955 to 1,675 lines (-280).

Part of #1433 (carve mega-modules into focused submodules).
